introduce tableManager.
